Convert 'Schedule a call' CTA to a reusable component.
