+++ client/a8c-for-agencies/sections/overview/sidebar/growth-accelerator/index.tsx
@@ -1,40 +1,22 @@
 import { Card } from '@automattic/components';
-import { Button, Icon } from '@wordpress/components';
-import { external } from '@wordpress/icons';
+import { Button } from '@wordpress/components';
 import { useTranslate } from 'i18n-calypso';
 import { useCallback } from 'react';
-import useFetchScheduleCallLink from 'calypso/a8c-for-agencies/data/agencies/use-fetch-schedule-call-link';
 import { useDispatch, useSelector } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import { savePreference } from 'calypso/state/preferences/actions';
 import { getPreference } from 'calypso/state/preferences/selectors';
+import OverviewSidebarGrowthAcceleratorCta from './cta';
 
 import './style.scss';
 
-const GROWTH_ACCELERATOR_REQUESTED_PREFERENCE = 'a4a_growth_accelerator_requested';
-const GROWTH_ACCELERATOR_DISMISSED_PREFERENCE = 'a4a_growth_accelerator_dismissed';
+export const GROWTH_ACCELERATOR_REQUESTED_PREFERENCE = 'a4a_growth_accelerator_requested';
+export const GROWTH_ACCELERATOR_DISMISSED_PREFERENCE = 'a4a_growth_accelerator_dismissed';
 
 export default function OverviewSidebarGrowthAccelerator() {
 	const translate = useTranslate();
 
 	const dispatch = useDispatch();
-
-	const { refetch: fetchScheduleCallLink, isFetching: isFetchingScheduleCallLink } =
-		useFetchScheduleCallLink();
-
-	const onRequestCallClick = useCallback( () => {
-		dispatch( recordTracksEvent( 'calypso_a4a_overview_growth_accelerator_schedule_call_click' ) );
-		dispatch( savePreference( GROWTH_ACCELERATOR_REQUESTED_PREFERENCE, true ) );
-
-		fetchScheduleCallLink().then( ( result ) => {
-			window.open(
-				result.data
-					? result.data
-					: 'https://meetings.hubspot.com/automattic-for-agencies/discovery-meeting',
-				'_blank'
-			);
-		} );
-	}, [ dispatch, fetchScheduleCallLink ] );
 
 	const onNotInterestedClick = useCallback( () => {
 		dispatch( recordTracksEvent( 'calypso_a4a_overview_growth_accelerator_not_interested_click' ) );
@@ -66,16 +48,7 @@ export default function OverviewSidebarGrowthAccelerator() {
 			</p>
 
 			<div className="overview__growth-accelerator-footer">
-				<Button
-					className="overview__growth-accelerator-footer-schedule-call"
-					variant="primary"
-					disabled={ isFetchingScheduleCallLink }
-					isBusy={ isFetchingScheduleCallLink }
-					onClick={ onRequestCallClick }
-				>
-					{ translate( 'Schedule a call' ) }
-					<Icon icon={ external } size={ 16 } />
-				</Button>
+				<OverviewSidebarGrowthAcceleratorCta />
 				<Button
 					className="overview__growth-accelerator-footer-not-interested"
 					variant="link"
